chore: refactor profiler table
